+++ rnalysis_app.py
@@ -1,7 +1,8 @@
+import asyncio
 from multiprocessing import freeze_support
 
 from rnalysis.gui import gui
 
 if __name__ == '__main__':
     freeze_support()
-    gui.run()
+    asyncio.run(gui.run())