+++ HISTORY.rst
@@ -21,6 +21,7 @@ Changed
 * Pipelines that contain consecutive clustering/splitting functions will now return their outputs in a clearer format.
 * Enrichment bar-plots should now adjust the x-axis limits more tightly to fit the displayed data.
 * Improved clarity of automatic titles in enrichment plots.
+* Download/update speed of tutorial videos has improved significantly.
 
 Fixed
 ******